requirements.txt
Django==4.2.4

Scrapy==2.10.1
scrapy-djangoitem==1.1.1
requests==2.31.0

python-dotenv==1.0.0

python-telegram-bot==20.4

pytest==7.0.1
pytest-cov==3.0.0
pytest-sugar==0.9.4

black==23.7.0
flake8==6.1.0
pre-commit
